Refactor - Renomeia e move classes de pacotes.
