+++ src/main/java/com/github/savitoh/centralerroapi/evento/tipologlevel/TipoLogLevel.java
@@ -1,4 +1,4 @@
-package com.github.savitoh.centralerroapi.evento;
+package com.github.savitoh.centralerroapi.evento.tipologlevel;
 
 import java.util.Optional;
 import java.util.stream.Stream;
