+++ src/main/java/com/github/savitoh/centralerroapi/seguranca/config/SecurityConfig.java
@@ -31,7 +31,7 @@ import java.io.IOException;
 @Order(2)
 @Configuration
 @EnableWebSecurity
-public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Autowired
     private UserPrincipalDetailsService userPrincipalDetailsService;
